+++ RELEASENOTES.md
@@ -64,6 +64,13 @@
         being enabled). Any changes made to the Player outside of the
         observation period are now picked up
         ([#2313](https://github.com/androidx/media/issues/2313)).
+    *   Add support for ExoPlayer's scrubbing mode to `PlayerControlView`. When
+        enabled, this puts the player into scrubbing mode when the user starts
+        dragging the scrubber bar, issues a `player.seekTo` call for every
+        movement, and then exits scrubbing mode when the touch is lifted from
+        the screen. This integration can be enabled with either
+        `time_bar_scrubbing_enabled = true` in XML or the
+        `setTimeBarScrubbingEnabled(boolean)` method from Java/Kotlin.
 *   Downloads:
     *   Add partial download support for progressive streams. Apps can prepare a
         progressive stream with `DownloadHelper`, and request a